+++ app/TypiCMS/Modules/Events/Models/Event.php
@@ -60,24 +60,4 @@ class Event extends Base
      * @var string
      */
     public $route = 'events';
-
-    /**
-     * Transform start_date in Carbon object
-     *
-     * @param string $value date string
-     */
-    public function setStartDateAttribute($value)
-    {
-        $this->attributes['start_date'] = Carbon::parse($value);
-    }
-
-    /**
-     * Transform end_date in Carbon object
-     *
-     * @param string $value date string
-     */
-    public function setEndDateAttribute($value)
-    {
-        $this->attributes['end_date'] = Carbon::parse($value);
-    }
 }