+++ app/database/migrations/2013_08_31_000501_create_events_table.php
@@ -19,8 +19,6 @@ class CreateEventsTable extends Migration
 
             $table->timestamp('start_date');
             $table->timestamp('end_date')->nullable();
-            $table->string('start_time')->nullable();
-            $table->string('end_time')->nullable();
             $table->string('image')->nullable();
 
             $table->timestamps();